Allow PCI BARs to be more than 16-bits in length (#3032)
This change makes it possible to have virtual PCI BAR of length greater
than 2^16. This is necessary for connecting simulated chips with larger
BARs, or for adding virtual devices with a whole lot of doorbells, or
anything else which requires a large BAR.

Most of the virtual devices are updated to just use 64-bit offsets in
their read/write functions. The virtio stuff is stubbed back to 16-bit
for convenience.

+++ vm/devices/pci/pci_core/src/spec.rs
@@ -465,7 +465,7 @@ pub mod caps {
 
         open_enum::open_enum! {
             /// Offsets into a single MSI-X Table Entry
-            pub enum MsixTableEntryIdx: u16 {
+            pub enum MsixTableEntryIdx: u64 {
                 MSG_ADDR_LO = 0x00,
                 MSG_ADDR_HI = 0x04,
                 MSG_DATA    = 0x08,
